0.15
- Se organiza código, se emprolijan funciones, se eliminan comentarios que no sirven.
