+++ app/src/main/java/com/example/espncito/ui/main/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.example.espncito.ui.main
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
@@ -9,7 +8,6 @@ import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.example.espncito.databinding.ActivityMainBinding
 import com.example.espncito.ui.viewByLeague.ViewByLeagueFragment
-import com.example.espncito.ui.viewByTeam.ViewByTeamActivity
 
 class MainActivity : AppCompatActivity() {
 
@@ -22,11 +20,6 @@ class MainActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         setupViewPager()
-
-        //Prueba para ir al activity de viewByTeam, eliminar luego de pruebas.
-/*        val intent = Intent(this, ViewByTeamActivity::class.java)
-        startActivity(intent)
-        finish()*/
     }
 
     private fun setupViewPager() {
